requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==4.1.7
django-cache-url==3.4.4
django-ckeditor==6.5.1
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-mail-templated==2.6.5
django-modeltranslation==0.18.9
django-rosetta==0.9.8
environs==9.5.0
idna==3.4
jdatetime==4.1.1
marshmallow==3.19.0
packaging==23.0
Pillow==9.5.0
polib==1.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.2
urllib3==1.26.15
